Update to 367.18.
Make a soft dependency for nvidia-uvm as adding the module loading to
/usr/lib/modules-load.d/nvidia-uvm.conf for systemd consumption, makes the
configuration file to be added to the initrd but not the module, throwing an
error on plymouth about not being able to find the module.
Ref: /usr/lib/dracut/modules.d/00systemd/module-setup.sh

Even adding the module is not the correct thing, as we don't want it to be
included in the initrd, so use this configuration file to specify the
dependency.