allow for pre-downloaded .run files
If the expected .run file is already present, just make a symlink to it
with the name we expect to work with. Otherwise download the file and
rename it to our desired working name.

Since the rm at the end of create_tarball just deletes the working name,
if the .run file already existed, only the symlink will be deleted and
the .run file will remain.