Update to 396.18, and build i386 driver-libs from x86_64 sources
Now that NVIDIA has decided not to support 32-bit systems any further,
the i386 driver libs must be pulled from the 64-bit driver package.

Add a `--without driver` option to the spec file to support that. When
it is enabled, none of the subpackages that require files not included
in the 32-bit compat libs folder are built.

To make it work, the subpackages need to be changed a bit:

- Add the nvidia-driver-headers subpackage. Move everything in
/usr/include to it, and make it a dependency. This is needed since the
headers are not included in the 32-bit compat libs, and hence can only
be built from the full 64-bit driver.
- Move libnvidia-cfg.so.* out of the libs package. It is only used by
nvidia-xconfig, and is not included in the 32-bit compat libs.
- Move the EGL vendor configurations to the main driver package, as it
is not present in the 32-bit compat libs.
- Add libnvidia-glvkspirv.so.* to the libs package, as it was introduced
in 396.18