Rename libnvidia-vulkan-producer.so library (#128)
The runfile provides libnvidia-vulkan-producer.so.515.43.04 without a SONAME, so ldconfig -vn . does not catch the need for a symlink.

The driver actually looks for libnvidia-vulkan-producer.so without the version for Vulkan under Wayland. You can actually see the dlopens with the following command:

libnvidia-vulkan-producer.so
libnvidia-vulkan-producer.so

The error can be triggered with "vulkaninfo" under a Wayland session.

Adding a simple symlink does not work as then RPM adds a requirement to "libnvidia-vulkan-producer.so()(64bit)", which is then not satisfied by the package.

It seems the only solution is renaming the library until a fix is in place.